Rename course_data app to data_aggregator
