+++ README.md
@@ -1,3 +1,3 @@
 # canvas-analytics
 * analytics/ - Builds reports about UW Canvas usage.
-* course_data/ - DB loader and job manager for UW Canvas analytics data.
+* data_aggregator/ - DB loader and job manager for UW Canvas analytics data.
