--- course_data/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class CourseDataConfig(AppConfig):
-    name = 'course_data'
