+++ data_aggregator/VERSION
