+++ data_aggregator/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DataAggregatorConfig(AppConfig):
+    name = 'data_aggregator'
