+++ data_aggregator/management/commands/create_participation_jobs.py
@@ -1,7 +1,7 @@
 
 from django.core.management.base import BaseCommand
-from course_data.logger import Logger
-from course_data.models import Course, Job, JobType
+from data_aggregator.logger import Logger
+from data_aggregator.models import Course, Job, JobType
 from django.db.models import Q
 from django.utils import timezone
 from datetime import timedelta
@@ -10,7 +10,7 @@ from uw_sws.term import get_current_term
 
 class Command(BaseCommand):
 
-    help = ("Creates assignment jobs for active courses in current term.")
+    help = ("Creates participation jobs for active courses in current term.")
 
     def add_arguments(self, parser):
         parser.add_argument("--log_file",
@@ -25,8 +25,8 @@ class Command(BaseCommand):
         """
         self.logger = Logger(logpath=options["log_file"])
 
-        # get assignment job type
-        assignment_type, _ = JobType.objects.get_or_create(type="assignment")
+        # get participation job type
+        partic_type, _ = JobType.objects.get_or_create(type="participation")
 
         # set target bounds from monday to sunday (work week)
         today = timezone.now().date()
@@ -42,18 +42,18 @@ class Command(BaseCommand):
         if course_count == 0:
             self.logger.info(
                 f'No active courses in term {sws_term.canvas_sis_id()} to '
-                f'create assignment jobs for.')
+                f'create participation jobs for.')
         else:
             for course in courses:
-                # create assignment jobs
+                # create participation jobs
                 self.logger.info(
-                    f"Adding assignment jobs for course "
+                    f"Adding participation jobs for course "
                     f"{course.canvas_course_id}")
                 job = Job()
-                job.type = assignment_type
+                job.type = partic_type
                 job.target_date_start = target_date_start
                 job.target_date_end = target_date_end
                 job.context = {'canvas_course_id': course.canvas_course_id}
                 job.save()
                 jobs_count += 1
-        self.logger.info(f'Created {jobs_count} assignment jobs.')
+        self.logger.info(f'Created {jobs_count} participation jobs.')
