+++ data_aggregator/static/css/course_data/custom.scss
@@ -1,0 +1,6 @@
+// myuw theme variables
+@import "./variables.scss";
+
+// bootstrap + bootstrap-vue
+@import '~bootstrap';
+@import '~bootstrap-vue';
