+++ data_aggregator/static/vue/base.js
