+++ data_aggregator/static/vue/components/home/jobs-range-picker.vue
