+++ data_aggregator/static/vue/components/home/jobs-table.vue
