+++ data_aggregator/static/vue/mixins/data_mixin.js
