+++ data_aggregator/static/vue/mixins/utilities_mixin.js
