+++ data_aggregator/static/vue/vuex/store/home_store.js
