+++ data_aggregator/templates/home.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 
 {% block body_content %}
-<h3 class="mb-4">Canvas Analytics Jobs</h3>
+<h3 class="mb-4">Canvas Data Aggregator Jobs</h3>
 <jobs-filter></jobs-filter>
 <jobs-table :jobs="filteredJobs" :selected-jobs="selectedJobs"></jobs-table>
 {% endblock body_content %}
