+++ docker/settings.py
@@ -2,12 +2,12 @@ from .base_settings import *
 import os
 
 INSTALLED_APPS += [
-    'course_data.apps.CourseDataConfig',
+    'data_aggregator.apps.DataAggregatorConfig',
 ]
 
 if os.getenv('ENV') == 'localdev':
     DEBUG = True
-    COURSE_DATA_ADMIN_GROUP = 'u_test_group'
+    DATA_AGGREGATOR_ADMIN_GROUP = 'u_test_group'
     RESTCLIENTS_DAO_CACHE_CLASS = None
 else:
-    COURSE_DATA_ADMIN_GROUP = os.getenv('ADMIN_GROUP', '')
+    DATA_AGGREGATOR__ADMIN_GROUP = os.getenv('ADMIN_GROUP', '')
